Read Data from GeneBank Files

This function reads data from a single GeneBank file or directorty with GenBank files. It allows selective extraction of information by specifying sections and features.

Usage

read_gbk(path, sections = NULL, features = NULL, origin = TRUE)

Arguments

path

A string representing the file path to the target GeneBank (.gbk) file or directory.

sections

An optional vector of strings representing the names of specific sections within the GeneBank file to extract (e.g., "LOCUS", "DEFINITION", "ACCESSION", "VERSION"). If `NULL` (the default), the function extracts all available sections.

features

An optional vector of strings indicating specific feature types to extract from the FEATURES section of the GeneBank file (e.g., "CDS", "gene", "mRNA"). If `NULL` (the default), the function extracts all feature types present in the FEATURES section.

origin

A boolean flag; when set to `TRUE` (the default), the origin sequence data is included in the output.

Value

A list containing the contents of the specified sections and features of the GeneBank file. Each section and feature is returned as a separate list element.

Examples

if (FALSE) {
# Read all data from a GeneBank file
gbk_data <- read_gbk("path/to/genebank_file.gbk")

# Read all data from a directory of GeneBank files
gbk_data <- read_gbk("path/to/genebank/directory")

# Read only specific sections from a GeneBank file
gbk_data <- read_gbk(
  "path/to/genebank_file.gbk",
  sections = c("LOCUS", "DEFINITION")
)

# Read specific features from the FEATURES section of a GeneBank file
gbk_data <- read_gbk("path/to/genebank_file.gbk", features = c("gene", "CDS"))

# Read data without the origin sequence
gbk_data <- read_gbk("path/to/genebank_file.gbk", origin = FALSE)
}